pyproject.toml
[project]
name = "mymadr"
version = "0.1.0"
description = ""
romancistas = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.120.0,<0.121.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "sqlalchemy[asynio] (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "aiosqlite (>=0.21.0,<0.22.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.14.2,<0.15.0)",
    "pytest-asyncio (>=1.2.0,<2.0.0)"
]

[tool.ruff]
line-length = 79

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope= "function"

[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev mymadr/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=mymadr -vv'
post_test = 'coverage html'